fix: Moved view context menu content to a different reusable component, added max length validation to view title and improved view title renamed UX
